requirements.txt
alembic==1.9.1
anyio==3.6.2
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.88.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
importlib-metadata==5.2.0
importlib-resources==5.10.1
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2-binary==2.9.5
pyasn1==0.4.8
pydantic==1.10.2
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.45
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
zipp==3.11.0
pydantic-settings==2.0.0
celery==5.3.4
redis==5.0.1
flower==2.0.1
